requirements.txt
# Google ADK
google-adk>=1.13.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Google Cloud Functions
functions-framework>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# MCP
mcp>=1.8.0

# FastAPI (ถ้าใช้)
fastapi>=0.100.0
uvicorn>=0.20.0

# Other dependencies
pydantic>=2.0.0
